+++ tests/testthat/test-dates.R
@@ -13,15 +13,15 @@ context("Dates")
 
 test_that("1900 and 1904 dates are parsed correctly", {
   expect_equal(tidy_xlsx("./1900.xlsx")$data$Sheet1$date[1], as.POSIXct("1998-07-05", tz = "UTC"))
+  expect_equal(tidy_xlsx("./1900.xlsx")$data$Sheet1$date[3], as.POSIXct("1900-02-28", tz = "UTC"))
   expect_equal(tidy_xlsx("./1904.xlsx")$data$Sheet1$date[1], as.POSIXct("1998-07-05", tz = "UTC"))
-  expect_warning(expect_equal(tidy_xlsx("./openxlsx1900.xlsx")$data$Sheet1$date[2],
-                              as.POSIXct("1899-12-31", tz = "UTC")),
-                 "Dates before 1 March 1900 are off by one, due to Excel's famous bug.")
-})
-
-test_that("A warning is given for 1900-system dates before 1 March 1900", {
-  expect_warning(tidy_xlsx("./date-bug.xlsx"), "Dates before 1 March 1900 are off by one, due to Excel's famous bug.")
-  expect_warning(tidy_xlsx("./openxlsx1900.xlsx"), "Dates before 1 March 1900 are off by one, due to Excel's famous bug.")
+  # This is the only way to construct an NA POSIXct with timezone
+  NA_POSIXct <- as.POSIXct(NA)
+  attr(NA_POSIXct, "tzone") <- "UTC"
+  expect_warning(expect_equal(tidy_xlsx("./1900-02-29.xlsx")$data$Sheet1$date[1], NA_POSIXct),
+                 "NA inserted for impossible 1900-02-29 datetime")
+  expect_equal(tidy_xlsx("./openxlsx1900.xlsx")$data$Sheet1$date[2],
+               as.POSIXct("1900-01-01", tz = "UTC"))
 })
 
 test_that("Custom number formats using [Red] aren't detected as dates", {